+++ CHANGELOG.md
@@ -55,6 +55,7 @@
 - SALE ORDER : Update in 'Quotations template' working process and view.
 - Business Project : Addition of new m2o field 'customerAddress'.
 - Base : Additon of new object 'MailingListMessage' along with views and Menu.
+- DataBackup : Relative date
 
 ## Improvements
 - SaleOrder: refactored 'Business Project' panel